requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
matplotlib>=3.8.0
reportlab>=4.0.0
faker>=19.0.0
openpyxl>=3.1.0
pandas>=2.1.0
gunicorn>=21.0.0
whitenoise>=6.6.0
